Rename source.typedtext package to source
