+++ scalding-commons/src/main/scala/com/twitter/scalding/commons/source/LzoTypedText.scala
@@ -1,11 +1,12 @@
-package com.twitter.scalding.commons.source.lzotypedtext
+package com.twitter.scalding.commons.source
 
 import cascading.scheme.Scheme
 import cascading.scheme.hadoop.{ TextDelimited => CHTextDelimited }
 import cascading.scheme.local.{ TextDelimited => CLTextDelimited }
 import com.twitter.elephantbird.cascading2.scheme.LzoTextDelimited
 import com.twitter.scalding._
-import com.twitter.scalding.source.typedtext._
+import com.twitter.scalding.source.TypedTextDelimited
+import com.twitter.scalding.source.TypedSep
 import com.twitter.scalding.typed.TypedSink
 
 object LzoTypedText {
