+++ scalding-core/src/main/scala/com/twitter/scalding/source/TypedText.scala
@@ -1,4 +1,4 @@
-package com.twitter.scalding.source.typedtext
+package com.twitter.scalding.source
 
 import cascading.scheme.Scheme
 import cascading.scheme.hadoop.{ TextDelimited => CHTextDelimited }
@@ -15,7 +15,7 @@ case class TypedSep(str: String) extends AnyVal
 object TypedText {
 
   private val TAB = TypedSep("\t")
-  private val ONE = TypedSep("\1")
+  private val ONE = TypedSep("\u0001")
   private val COMMA = TypedSep(",")
 
   def tsv[T: TypeDescriptor](path: String*): TypedTextDelimited[T] = new FixedTypedText[T](TAB, path: _*)
